Rename package to artstationapplication.
