+++ src/artstationapplication/Triangle.java
